backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
ollama==0.1.6
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.2.0
requests==2.31.0
mammoth==1.6.0